style: [#3] change code style
-enum VideoFormatCapacity를 별도 파일로 분리
-설명 주석을 수정
-func 파라미터명과 유형 사이 ":" 위치를 수정
