--- MoonCrystal/MoonCrystal/StorageManager.swift
@@ -1,8 +1,0 @@
-//
-//  StorageManager.swift
-//  MoonCrystal
-//
-//  Created by seonu kim on 7/29/24.
-//
-
-import Foundation
